+++ uniffi_bindgen/src/bindings/kotlin/gen_kotlin/mod.rs
@@ -439,7 +439,9 @@ impl KotlinCodeOracle {
             FfiType::ForeignBytes => "ForeignBytes.ByValue".to_string(),
             FfiType::Callback(name) => self.ffi_callback_name(name),
             FfiType::Struct(name) => self.ffi_struct_name(name),
-            FfiType::Reference(inner) => self.ffi_type_label_by_reference(inner),
+            FfiType::Reference(inner) | FfiType::MutReference(inner) => {
+                self.ffi_type_label_by_reference(inner)
+            }
             FfiType::VoidPointer => "Pointer".to_string(),
         }
     }